refactor: move DOM manipulation into builder.js
