+++ app.js
@@ -1,3 +1,5 @@
+import { buildNextHoliday, buildNextLongWeekend, buildSubsequentHoliday } from "./builder.js";
+
 let indexForHolidays = 0;
 let indexForLongWeekends = 0;
 document.addEventListener("DOMContentLoaded", getHolidays());
@@ -9,16 +11,8 @@ async function getHolidays(){
 
     const currentDate = new Date();
     const nextHoliday = holidays.holidays.find(holiday => new Date(holiday.date) > currentDate);
-    let dateOptions = { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' };
 
-    let mainContainer = document.getElementById('nextHolidayContainer');
-    mainContainer.textContent = '';
-    let holidayName = document.createElement('p');
-    let holidayDate = document.createElement('p');
-    holidayName.textContent = nextHoliday.name;
-    holidayDate.textContent = new Date(nextHoliday.date).toLocaleDateString('en-US', dateOptions);
-    mainContainer.appendChild(holidayName);
-    mainContainer.appendChild(holidayDate);
+    buildNextHoliday(nextHoliday);
     indexForHolidays = 0;
     indexForLongWeekends = 0;
     getNextLongWeekend();
@@ -34,15 +28,7 @@ async function getSubsequentHoliday(){
     const nextHolidays = holidays.holidays.filter(holiday => new Date(holiday.date) > currentDate)
     console.log(nextHolidays[indexForHolidays])
     indexForHolidays++;
-    let dateOptions = { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' };
-    let mainContainer = document.getElementById('nextHolidayContainer');
-    mainContainer.textContent = '';
-    let holidayName = document.createElement('p');
-    let holidayDate = document.createElement('p');
-    holidayName.textContent = nextHolidays[indexForHolidays].name;
-    holidayDate.textContent = new Date(nextHolidays[indexForHolidays].date).toLocaleDateString('en-US', dateOptions);
-    mainContainer.appendChild(holidayName);
-    mainContainer.appendChild(holidayDate);
+    buildSubsequentHoliday(nextHolidays, indexForHolidays);
     countdownForHolidays(nextHolidays[indexForHolidays].date);
     return nextHolidays;
 }
@@ -54,15 +40,7 @@ async function getNextLongWeekend(){
     const currentDate = new Date();
     let nextLongWeekends = holidays.holidays.filter(holiday => holiday.longWeekend === true).filter(holiday => new Date(holiday.date) > currentDate);
 
-    let dateOptions = { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' };
-    let mainContainer = document.getElementById('nextLongWeekendContainer');
-    mainContainer.textContent = '';
-    let holidayName = document.createElement('p');
-    let holidayDate = document.createElement('p');
-    holidayName.textContent = nextLongWeekends[indexForLongWeekends].name;
-    holidayDate.textContent = new Date(nextLongWeekends[indexForLongWeekends].date).toLocaleDateString('en-US', dateOptions);
-    mainContainer.appendChild(holidayName);
-    mainContainer.appendChild(holidayDate);
+    buildNextLongWeekend(nextLongWeekends, indexForLongWeekends);
     countdownForLongWeekends(nextLongWeekends[indexForLongWeekends].date);
     indexForLongWeekends++;
 }
@@ -83,3 +61,8 @@ function countdownForLongWeekends(nextLongWeekendDate){
     let h2 = document.getElementById('h2');
     h2.textContent = `The next long weekend is on: ${dayDifference} days later`;
 }
+
+// NOTE: not a good practice, but serves the purpose now
+window.getHolidays = getHolidays;
+window.getSubsequentHoliday = getSubsequentHoliday;
+window.getNextLongWeekend = getNextLongWeekend;
