+++ index.html
@@ -7,7 +7,6 @@
     <title>Holidayssss</title>
 </head>
 <body>
-    <script src="app.js"></script>
     <h1 id="h1">The next holiday is on: </h1>
     <div id="nextHolidayContainer"></div>
     <button onclick="getHolidays()">Reset</button>
@@ -17,5 +16,6 @@
     <button onclick="getNextLongWeekend()">Show Me Next Long Weekend</button>
     <h2 id="h2">The next long weekend is on:</h2>
     <div id="nextLongWeekendContainer"></div>
+    <script type="module" src="app.js"></script>
 </body>
 </html>
